+++ src/signalfx_metadata.py
@@ -56,7 +56,7 @@ API_TOKENS = []
 TIMEOUT = 3
 POST_URLS = []
 DEFAULT_POST_URL = "https://ingest.signalfx.com/v1/collectd"
-VERSION = "0.0.19"
+VERSION = "0.0.20"
 NOTIFY_LEVEL = -1
 HOST_TYPE_INSTANCE = "host-meta-data"
 TOP_TYPE_INSTANCE = "top-info"
@@ -275,7 +275,7 @@ class MemoryUtilization(Utilization):
             m = self.metrics[t]
             if len(m) == self.size:
                 total = sum(c[0] for c in m.values())
-                used = total - m["memory.free"][0]
+                used = m["memory.used"][0]
                 self.emit_utilization(t, used, total,
                                       "memory.utilization", obj=m)
             else:
@@ -1156,13 +1156,14 @@ def put_val(plugin_instance, type_instance, val, plugin=PLUGIN_NAME, t=0.0,
                 type_instance, INTERVAL, val[0]))
     except TypeError:
         global UTILIZATION
-        UTILIZATION = False
-        log("ERROR: Utilization features have been disabled because TypesDB" +
-            " hasn't been specified")
-        log("To use the utilization features of this plugin, please update" +
-            " the top of your config to include" +
-            " 'TypesDB \"/opt/signalfx-collectd-plugin/types.db.plugin\"'")
-        raise
+        if UTILIZATION:
+            UTILIZATION = False
+            log("ERROR: Utilization features have been disabled because " +
+                "TypesDB hasn't been specified")
+            log(
+                "To use the utilization features of this plugin, please " +
+                "update the top of your config to include 'TypesDB " +
+                "\"/opt/signalfx-collectd-plugin/types.db.plugin\"'")
 
 
 def get_uptime():