+++ CHANGELOG.md
@@ -1,5 +1,11 @@
-0.0.19 / 2016-03-25
+0.0.20 / 2016-03-31
 ===================
+
+* change memory.utilization to not include cached, bufferred or slab but just
+  be used/total
+* fix repetative logging when types.db.plugin isn't found
+
+0.0.19 / 2016-03-25
 
 * add ability for multiple endpoints